backend/requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Django==3.1.4
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
pip==20.2.4
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.4
requests==2.25.0
setuptools==50.3.2
sqlparse==0.4.1
urllib3==1.26.2
waitress==1.4.4
wheel==0.35.1
whitenoise==5.2.0
